remove part of duplicate code
