+++ plugins/core/jetbrains-community/src/software/aws/toolkits/jetbrains/core/webview/LoginBrowser.kt
@@ -3,12 +3,17 @@
 
 package software.aws.toolkits.jetbrains.core.webview
 
+import com.intellij.openapi.application.runInEdt
 import com.intellij.openapi.project.Project
+import com.intellij.ui.JBColor
 import com.intellij.ui.jcef.JBCefBrowserBase
 import com.intellij.ui.jcef.JBCefJSQuery
 import kotlinx.coroutines.launch
 import software.aws.toolkits.core.credentials.ToolkitBearerTokenProvider
+import software.aws.toolkits.core.region.AwsRegion
 import software.aws.toolkits.jetbrains.core.coroutines.projectCoroutineScope
+import software.aws.toolkits.jetbrains.core.credentials.AwsBearerTokenConnection
+import software.aws.toolkits.jetbrains.core.credentials.Login
 import software.aws.toolkits.jetbrains.core.credentials.ManagedBearerSsoConnection
 import software.aws.toolkits.jetbrains.core.credentials.ToolkitAuthManager
 import software.aws.toolkits.jetbrains.core.credentials.ToolkitConnection
@@ -20,17 +25,21 @@ import software.aws.toolkits.jetbrains.utils.pollFor
 import software.aws.toolkits.telemetry.FeatureId
 import java.util.function.Function
 
-data class BrowserState(val feature: FeatureId, val browserCancellable: Boolean = false)
+data class BrowserState(val feature: FeatureId, val browserCancellable: Boolean = false, val requireReauth: Boolean = false)
 
 abstract class LoginBrowser(
     private val project: Project,
     val domain: String,
+    val webScriptUri: String
 ) {
     abstract val jcefBrowser: JBCefBrowserBase
     abstract val query: JBCefJSQuery
     abstract val handler: Function<String, JBCefJSQuery.Response>
 
     protected var currentAuthorization: PendingAuthorization? = null
+
+    protected data class BearerConnectionSelectionSettings(val currentSelection: AwsBearerTokenConnection, val onChange: (AwsBearerTokenConnection) -> Unit)
+    protected val selectionSettings = mutableMapOf<String, BearerConnectionSelectionSettings>()
 
     // TODO: figure out a better way to do this UI update
     protected val onPendingProfile: (InteractiveBearerTokenProvider) -> Unit = { provider ->
@@ -78,11 +87,68 @@ abstract class LoginBrowser(
         executeJS("window.ideClient.reset()")
     }
 
+    open fun loginBuilderId(scopes: List<String>) {
+        runInEdt {
+            Login.BuilderId(scopes, onPendingAwsId).loginBuilderId(project)
+            // TODO: telemetry
+        }
+    }
+
+    open fun loginIdC(profileName: String, url: String, region: AwsRegion, scopes: List<String>) {
+        val onError: (String) -> Unit = { _ ->
+            // TODO: telemetry
+        }
+        runInEdt {
+            Login.IdC(profileName, url, region, scopes, onPendingProfile, onError).loginIdc(project)
+            // TODO: telemetry
+        }
+    }
+
+    open fun loginIAM(profileName: String, accessKey: String, secretKey: String) {
+        // TODO: telemetry, callbacks
+        runInEdt {
+            Login.LongLivedIAM(
+                profileName,
+                accessKey,
+                secretKey
+            ).loginIAM(project, {}, {}, {})
+        }
+    }
+
     protected fun loadWebView() {
         jcefBrowser.loadHTML(getWebviewHTML())
     }
 
-    protected abstract fun getWebviewHTML(): String
+    protected fun getWebviewHTML(): String {
+        val colorMode = if (JBColor.isBright()) "jb-light" else "jb-dark"
+        val postMessageToJavaJsCode = query.inject("JSON.stringify(message)")
+
+        val jsScripts = """
+            <script>
+                (function() {
+                    window.ideApi = {
+                     postMessage: message => {
+                         $postMessageToJavaJsCode
+                     }
+                };
+                }())
+            </script>
+            <script type="text/javascript" src="$webScriptUri"></script>
+        """.trimIndent()
+
+        return """
+            <!DOCTYPE html>
+            <html>
+                <head>
+                    <title>AWS Q</title>
+                </head>
+                <body class="$colorMode">
+                    <div id="app"></div>
+                    $jsScripts
+                </body>
+            </html>
+        """.trimIndent()
+    }
 
     protected suspend fun pollForConnection(connectionId: String): ToolkitConnection? = pollFor {
         ToolkitAuthManager.getInstance().getConnection(connectionId)
