+++ plugins/toolkit/jetbrains-core/src/software/aws/toolkits/jetbrains/core/explorer/AwsToolkitExplorerFactory.kt
@@ -32,6 +32,7 @@ import software.aws.toolkits.jetbrains.core.explorer.webview.ToolkitWebviewPanel
 import software.aws.toolkits.jetbrains.core.help.HelpIds
 import software.aws.toolkits.jetbrains.core.webview.BrowserState
 import software.aws.toolkits.jetbrains.utils.actions.OpenBrowserAction
+import software.aws.toolkits.jetbrains.utils.inspectExistingConnection
 import software.aws.toolkits.resources.message
 import software.aws.toolkits.telemetry.FeatureId
 
@@ -169,34 +170,6 @@ class AwsToolkitExplorerFactory : ToolWindowFactory, DumbAware {
             contentManager.addContent(myContent)
         }
     }
-
-    private fun inspectExistingConnection(project: Project): Boolean =
-        ToolkitConnectionManager.getInstance(project).let {
-            if (CredentialManager.getInstance().getCredentialIdentifiers().isNotEmpty()) {
-                LOG.debug { "inspecting existing connection and found IAM credentials" }
-                return@let true
-            }
-
-            val conn = it.activeConnection()
-            val hasIdCRoleAccess = if (conn is AwsBearerTokenConnection) {
-                conn.scopes.contains(IDENTITY_CENTER_ROLE_ACCESS_SCOPE)
-            } else {
-                false
-            }
-
-            if (hasIdCRoleAccess) {
-                LOG.debug { "inspecting existing connection and found bearer connections with IdCRoleAccess scope" }
-                return@let true
-            }
-
-            val isCodecatalystConn = it.activeConnectionForFeature(CodeCatalystConnection.getInstance()) != null
-            if (isCodecatalystConn) {
-                LOG.debug { "inspecting existing connection and found active Codecatalyst connection" }
-                return@let true
-            }
-
-            return@let false
-        }
 
     companion object {
         private val LOG = getLogger<AwsToolkitExplorerFactory>()
